Bu_online/requirements.txt
aiogram==2.25.1
python-dotenv==1.0.0
requests==2.31.0
python-docx==0.8.11
pandas==2.0.3
openpyxl==3.1.2
apscheduler==3.10.1
python-dateutil==2.8.2